Use key events instead of <input> elements
Instead of using `<input maxlength="1">` as the cell input, we now just
use the keyDown event to set the value of cells. As part of this, the
backspace key is now used to clear cells and its default behaviour
(i.e., navigating backwards in history) is prevented.

Additionally, cells can now be focused. A focused cell is indicated with
a highlighted background. The keyDown event affects the value of the
focused cell (assuming it is not a locked cell).