Resize the Sudoku grid responsively
The Sudoku grid now resizes responsively with respect to the size of the
viewport, using the minimum of the width and height to determine its
dimentions. Additionally, a banner has been added and the grid now uses
purple (to match the new banner) for its input colour instead of blue.