Normalize cell state
Previously a cell in the grid was represented by:

```
interface Cell {
    readonly value: Value; // 1 .. 9 or null
    readonly locked: boolean;
}
```

and a `Sudoku` was a two-dimensional array of `Cell`s. Now, we represent
a Sudoku as a two-dimensional array of `Value`s and a two-dimensional
array of the lock states. Because of this change, to update a cell we no
longer have to create a new JS object for the cell, the row, and the
grid. Instead, we just create a new grid and row and replace the cell
with a new value, saving an allocation.