Simplify click/drag event handling
Instead of adding a click handler to the document (which, due to how
React proxies events, requires `Event#stopImmediatePropagation` to
prevent), there is now a top-level `SudokuMonster` component with a
`mouseup` handler. This handler takes the role of both the old top-level
`click` handler and the `mouseup` handlers of `Cell`s.

This also fixes a bug where dragging off the grid and letting go of the
mouse button would still result in focusing of cells when mousing over
them.